completed

+++ src/main/java/com/raaji/eventmanagement/repository/EventRepository.java
@@ -2,5 +2,5 @@ package com.raaji.eventmanagement.repository;
 import com.raaji.eventmanagement.entity.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
